Refactor folder structure of python pipeline. Use pyproject.toml instead of requirements.txt and setup.py. Move all packages into src folder.
