+++ pipeline/.env-example
@@ -1,0 +1,1 @@
+DAGSTER_HOME={absolute-path-to-repo}/yd_dagster_pipeline/data/.dagster
