+++ pipeline/pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "year-in-data-pipeline"
+version = "0.1.0"
+description = "Dagster pipelien with an api layer."
+authors = [{ name = "Aebel Shajan", email = "aebelshajan.projects@gmail.com" }]
+requires-python = ">=3.10,<3.11" # Due to dagster not playing nice with updated versions of python
+
+dependencies = [
+    "dagster",
+    "pandera",
+    "pandas",
+    "beautifulsoup4",    # For youtube html
+    "lxml",              # For youtube html
+]
+
+[project.optional-dependencies]
+dev = [
+    "dagster-webserver",
+    "pytest",
+    "ipykernel",
+]
+
+[tool.setuptools]
+package-dir = { "" = "src" }
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.dagster]
+module_name = ".definitions"
+code_location_name = "yd_dagster_pipeline"
