--- pipeline/requirements.txt
@@ -1,36 +1,0 @@
-annotated-types==0.7.0
-beautifulsoup4==4.13.4
-black==25.1.0
-certifi==2025.4.26
-charset-normalizer==3.4.2
-click==8.1.8
-colorama==0.4.6
-dotenv==0.9.9
-filelock==3.18.0
-gdown==5.2.0
-idna==3.10
-mypy_extensions==1.1.0
-numpy==2.2.5
-packaging==25.0
-pandas==2.2.3
-pandera==0.23.1
-pathspec==0.12.1
-platformdirs==4.3.8
-psutil==7.0.0
-pydantic==2.11.4
-pydantic_core==2.33.2
-PySocks==1.7.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.1.0
-pytz==2025.2
-requests==2.32.3
-six==1.17.0
-soupsieve==2.7
-toml==0.10.2
-tqdm==4.67.1
-typeguard==4.4.2
-typing-inspect==0.9.0
-typing-inspection==0.4.0
-typing_extensions==4.13.2
-tzdata==2025.2
-urllib3==2.4.0
