+++ pipeline/src/transformations/fitbit/schemas.py
