+++ pipeline/src/transformations/kindle/schemas.py
