+++ pipeline/src/transformations/strong/__init__.py
