+++ pipeline/src/transformations/strong/schemas.py
