+++ pipeline/src/transformations/utils/pipeline_stage.py
